[6.2.0][dev] 1.21 Fix nmsProxy
